v0.5.2 (PC->0.1.11)  -good counting during the test  -little changes with G1 GCode command

+++ Inc/a4988_stepstick.h
@@ -63,7 +63,7 @@ typedef enum MOTOR_RESET_FAZE{
 }MOTOR_RESET_FAZE;
 
 typedef struct RoundingErrorData{
-	double roundingMoveError;
+	int roundingMoveError;
 	double roundingSpeedError;
 	bool errMove;
 } RoundingErrorData;
